+++ CHANGELOG.md
@@ -7,6 +7,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Changed
+
+- Adjust filters - all directory filters also filter the directory itself (for when lstat/stating the directory).
+  Added `/Applications`
+
 ## 3.13.3
 
 ### Added